feat(field): Move annotation fields to its own tab
