+++ cypress/e2e/phase_4/annotation.cy.js
@@ -24,7 +24,7 @@ function loadFilmDataset() {
     );
     cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-    cy.findByRole('tab', { name: 'Semantics' }).click();
+    cy.findByRole('tab', { name: 'Annotations' }).click();
     cy.findByRole('checkbox', {
         name: 'This field can be annotated',
     }).click();
@@ -531,7 +531,7 @@ describe('Annotation', () => {
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByRole('checkbox', {
                 name: 'This field can be annotated',
             }).click();
@@ -680,7 +680,7 @@ describe('Annotation', () => {
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByRole('checkbox', {
                 name: 'This field can be annotated',
             }).click();
@@ -794,7 +794,7 @@ describe('Annotation', () => {
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByRole('checkbox', {
                 name: 'This field can be annotated',
             }).click();
@@ -922,7 +922,7 @@ describe('Annotation', () => {
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByLabelText('Annotation format').click();
 
             cy.findByRole('option', {
@@ -993,7 +993,7 @@ Revue`);
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByLabelText('Annotation format').click();
 
             cy.findByRole('option', {
@@ -1073,7 +1073,7 @@ Revue`);
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByRole('checkbox', {
                 name: 'This field can be annotated',
             }).click();
@@ -1203,7 +1203,7 @@ Revue`);
             );
             cy.findByRole('button', { name: 'edit-Liste des films' }).click();
 
-            cy.findByRole('tab', { name: 'Semantics' }).click();
+            cy.findByRole('tab', { name: 'Annotations' }).click();
             cy.findByRole('checkbox', {
                 name: 'This field can be annotated',
             }).click();
