+++ src/app/custom/translations.tsv
@@ -198,6 +198,7 @@
 "field_wizard_tab_display"	"Display"	"Affichage"
 "field_wizard_tab_search"	"Annotations"	"Annotations"
 "field_wizard_tab_semantic"	"Semantics"	"Sémantique"
+"field_wizard_tab_annotations"	"Annotations"	"Annotations"
 "field_wizard_warning_facet_message"	"The value to which the facet is applied is of type Object. Make sure to apply the necessary transformers"	"La valeur sur laquelle est appliquée la facette est de type Objet. Assurez-vous d'appliquer les transformations nécessaires"
 "previous"	"Previous"	"Précédent"
 "export_raw_dataset"	"Export raw dataset"	"Exporter le jeu de données"
@@ -1361,7 +1362,7 @@
 "annotation_deleted_by_admin"	"Some annotations have been deleted by the administrator"	"Des annotations ont été supprimées par l'administrateur"
 "annotation_sent_count"	"(%{smart_count} sent)||||(%{smart_count} sent)"	"(%{smart_count} envoyée)||||(%{smart_count} envoyées)"
 "annotation_disabled_tooltip"	"Annotations are disabled on this field"	"Les annotations sont désactivées sur ce champ"
-"field_annotation_format_list_supports_new_values"	"Ability to add a new value"	"Possibilité d'ajouter une nouvelle valeur"
+"field_annotation_format_list_supports_new_values"	"Allow contributors to suggest a new value"	"Permettre aux contributeurs de proposer une nouvelle valeur"
 "annotation_user_provided_value"	"New user provided value: %{value}"	"Nouvelle valeur fournie par l'utilisateur : %{value}"
 "autocomplete_open"	"Open"	"Ouvrir"
 "autocomplete_close"	"Close"	"Fermer"
